+++ src/Xteq5UserInterface/Xteq5UIConstant.cs
@@ -23,9 +23,9 @@ namespace Xteq5
             }
         }
 
-
         //MTH: These are static readonly fields BY PURPOSE (not CONST) to avoid that the compiler performs an copy+paste to the assembly of the consumer.
         //See [Const Strings - a very convenient way to shoot yourself in the foot](http://www.stum.de/2009/01/14/const-strings-a-very-convenient-way-to-shoot-yourself-in-the-foot/) by [Michael Stum](http://www.stum.de/)        
+
 
         public static readonly string ProjectURL = "https://github.com/texhex/Xteq5/wiki/";